migration: Add schema formatter
