+++ internal/database/migration/schemas/formatter.go
@@ -1,0 +1,5 @@
+package schemas
+
+type SchemaFormatter interface {
+	Format(schemaDescription SchemaDescription) string
+}
